requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
cloudinary==1.32.0
contourpy==1.0.7
crispy-bootstrap5==0.7
cycler==0.11.0
Django==4.1.7
django-allauth==0.52.0
django-crispy-forms==2.0
fonttools==4.39.0
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.24.2
oauthlib==3.2.2
pandas==1.5.3
Pillow==9.4.0
PyJWT==2.6.0
pyparsing==3.0.9
python3-openid==3.2.0
pytz==2022.7.1
requests-oauthlib==1.3.1
sqlparse==0.4.3
wordcloud==1.8.2.2
